Swappable light/dark solarized colorscheme.
Allows switching between the solarized dark and light colorschemes with
a keypress (F6 by default).

In config.def.h, this replaces colornames with the dark scheme and adds
a second colorname array (altcolornames) for the light scheme.

In st.c, accesses to colorname[i] are replaced with calls to
getcolorname(i), which chooses which colorname array to use based on
the usealtcolors flag. Pressing F6 toggles usealtcolors and triggers a
color load and redraw to swap the color scheme.